requirements.txt
spacy
PyPDF2
numpy
networkx
faiss-cpu
sentence-transformers
